pyproject.toml
[project]
package-mode = false
name = "gxp-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Elensito",email = "elensitopelaez@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "discord-py (>=2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
